Student Rader UPDATED
I have not made it so that everytime the users presess an button called find nearby students, the circles apeear on the screen with everyone of them spawning in random places everytime it loads. There is also another button there that I'm going to edit that will allow the user to send an email to that person and also try to intergrate it so that each student has an name that will apear at the side of the screen. Another addition may be to name diffrent sections for example the top right could be C&T Building and the top left be Media Factory