Rader *UPDATED*
I have manged to it to display the student names at the side of the canvas where you can also find them. However when tweaking the code the button to activate the canvas won't refresh once you have pressed it. I need to change this so you can press it and it updates automaticaly. Next I'm going to try to get the names changed just "student" to actual names and then when it displays the names at the side it can be clicked on like a button with an alert that will display an message.