Radar *UPDATED*
I have added some more functionality as now I have fixed the issue but also styled the name elements at the side of the page, added the alert function and got an button that makes all of the people go to one place. I have an issue at the moment where the alert doesnt show you the correct student when it displays an messege to say hello to them.

Another thing I have noticed is that my random colour no longer works as I have set each student an colour which might actualy be better than having them random but it does mean that John Readie is always on the radar 99% of the time.